+++ bot.py
@@ -17,8 +17,8 @@ from dotenv import load_dotenv
 if os.path.exists(".env"):
     load_dotenv()
 
-BOT_VERSION = "4.35.0"
-BOT_UPDATED = "2026-08-10"
+BOT_VERSION = "4.36.0"
+BOT_UPDATED = "2026-08-21"
 CHANGELOG_CHANNEL_ID = 1486967511839801414
 
 TOKEN = os.getenv("TOKEN")